+++ src/microsoft/winpro.c
@@ -31,7 +31,7 @@
 #include "../package.h"
 
 /* Suika2 HAL Implementaions */
-#include "d3drender.h"		/* Graphics HAL */
+#include "dx9render.h"		/* Graphics HAL */
 #include "dsound.h"			/* Sound HAL */
 #include "dsvideo.h"		/* Video HAL */
